requirements.txt
optimum
transformers
auto-gptq
torch
AutoAWQ
sentence-transformers
datasets
bitsandbytes
nvidia-ml-py3
pyarrow
tqdm
numpy
nltk
ctransformers
evaluate
rouge_score
lm-eval